% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{BarabasiAlbert}
\alias{BarabasiAlbert}
\title{Function to generate a Barabasi-Albert network}
\usage{
BarabasiAlbert(N, K)
}
\arguments{
\item{N}{the number of nodes}

\item{K}{the number of neighbors to which each node is connected so degree = 2*K}
}
\value{
An object of class CM (contact matrix)
}
\description{
Function to generate a Barabasi-Albert network
}
\examples{
cm3=BarabasiAlbert(200, 4)
}
