% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gm}
\alias{gm}
\title{Defoliated by gypsy moth each in northeast US 1975-2002.}
\format{A data frame with 1086 rows and 30 variables:
\describe{
  \item{UTMX}{UTM x-coordinates}
  \item{UTMY}{UTM y-coordinates}
 \item{d1975}{Defolitation in 1975}
 \item{d1976}{Defolitation in 1976}
 \item{d1977}{Defolitation in 1977}
 \item{d1978}{Defolitation in 1978}
 \item{d1979}{Defolitation in 1979}
  \item{d1980}{Defolitation in 1980}
  \item{d1981}{Defolitation in 1981}
  \item{d1982}{Defolitation in 1982}
  \item{d1983}{Defolitation in 1983}
  \item{d1984}{Defolitation in 1984}
  \item{d1985}{Defolitation in 1985}
  \item{d1986}{Defolitation in 1986}
  \item{d1987}{Defolitation in 1987}
  \item{d1988}{Defolitation in 1988}
  \item{d1989}{Defolitation in 1989}
  \item{d1990}{Defolitation in 1990}
  \item{d1991}{Defolitation in 1991}
  \item{d1992}{Defolitation in 1992}
  \item{d1993}{Defolitation in 1993}
  \item{d1994}{Defolitation in 1994}
  \item{d1995}{Defolitation in 1995}
  \item{d1996}{Defolitation in 1996}
  \item{d1997}{Defolitation in 1997}
  \item{d1998}{Defolitation in 1998}
  \item{d1999}{Defolitation in 1999}
  \item{d2000}{Defolitation in 2000}
  \item{d2001}{Defolitation in 2001}
  \item{d2002}{Defolitation in 2002}
}}
\source{
Bjornstad, O. N., Robinet, C., & Liebhold, A. M. (2010). Geographic variation in North American gypsy moth cycles: subharmonics, generalist predators, and spatial coupling. Ecology, 91(1), 106-118. doi.org/10.1890/08-1246.1.
}
\usage{
gm
}
\description{
A dataset containing the fraction of forest defoliated by the gypsy moth
in 20km x 20km pixels across northeast US in each year between 1975 and 2002.
}
\keyword{datasets}
