% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{NB}
\alias{NB}
\title{The Nicholson-Bailey model}
\usage{
NB(R, a, T = 100, H0 = 10, P0 = 1)
}
\arguments{
\item{R}{the host repruductive rate}

\item{a}{the parasite search efficiency}

\item{T}{the length of simulation (number of time-steps)}

\item{H0}{initial host numbers}

\item{P0}{initial parasitoid numbers}
}
\value{
A list of simulated Host and Parasitoid numbers
}
\description{
Function to simulate the Nicholson-Bailey Parasit-host model
}
\examples{
sim= NB(R=1.1,a=0.1)
}
