% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{plot.netSIR}
\alias{plot.netSIR}
\title{Function to plot a netSIR object}
\usage{
\method{plot}{netSIR}(x, ...)
}
\arguments{
\item{x}{an object of class netSIR}

\item{...}{other arguments}
}
\description{
Function to plot a netSIR object
}
\seealso{
\code{\link{netSIR}}
}
