% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{silene2}
\alias{silene2}
\title{Antler smut on wild campion.}
\format{A data frame with 876 rows and 5 variables:
\describe{
  \item{X}{road segment number}
  \item{lat}{latitude}
  \item{long}{longitude}
  \item{hmean}{number of healthy plants}
  \item{dmean}{number of diseased plants}
}}
\source{
Antonovics, J. 2004. Long-term study of a plant-pathogen metapopulation. In: Hanski, Ilkka, and Oscar E. Gaggiotti. Ecology, genetics, and evolution of metapopulations. Academic Press. doi.org/10.1371/journal.pone.0007359.
}
\usage{
silene2
}
\description{
Data on a fungal pathogen of the wild campion collected by Janis Antonovics
}
\keyword{datasets}
