% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{summary.netSIR}
\alias{summary.netSIR}
\title{Function to summarize a netSIR object}
\usage{
\method{summary}{netSIR}(object, ...)
}
\arguments{
\item{object}{an object of class netSIR}

\item{...}{other arguments}
}
\value{
A data-frame with the time series of suceptible, infected and recovered individuals
}
\description{
Function to summarize a netSIR object
}
\seealso{
\code{\link{netSIR}}
}
