% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuncs.R
\name{WattsStrogatz}
\alias{WattsStrogatz}
\title{Function to generate a Watts-Strogats network}
\usage{
WattsStrogatz(N, K, Prw)
}
\arguments{
\item{N}{the number of nodes}

\item{K}{the number of neighbors to which each node is connected so degree = 2xK}

\item{Prw}{the rewiring probability}
}
\value{
An object of class CM (contact matrix)
}
\description{
Function to generate a Watts-Strogats network
}
\examples{
cm2=WattsStrogatz(N=20, K=4, Prw=.3)
}
